+++ index.js
@@ -16,7 +16,7 @@ const client = new Client({
 });
 
 // --- Own Modules --- \\
-const wordle = require("./src/modules/wordle");
+const interactions = require("./src/modules/interactions");
 
 // --- Variables --- \\
 const token = process.env.token;
@@ -76,31 +76,7 @@ client.on('ready', _ => {
 });
 
 client.on('interactionCreate', async (interaction) => {
-  if(interaction.isCommand()) {
-    const { commandName, options } = interaction;
-
-    if(commandName === 'new'){
-      let id = options.getNumber('id');
-      wordle.newGame(id, interaction);
-    }else if(commandName === 'guess'){
-      let guess = options.getString('word');
-      const playNewBtn = new MessageActionRow()
-        .addComponents(
-          new MessageButton()
-          .setCustomId('playagain')
-          .setLabel("Play Again")
-          .setStyle('SUCCESS')
-        )
-      wordle.guess(guess, interaction, playNewBtn);
-    }else if(commandName === 'stats'){
-      let user = options.getUser('user');
-      wordle.stats(user, interaction);
-    }
-  }else if(interaction.isButton()){
-    if(interaction.customId == "playagain"){
-      wordle.newGame(null, interaction);
-    }
-  }
+  interactions.interactions(interaction);
 });
 
 // --- Bot Login --- \\