+++ src/modules/interactions.js
@@ -2,7 +2,7 @@ const { MessageSelectMenu, MessageActionRow } = require("discord.js");
 
 const wordle = require("./wordle.js");
 
-exports.interactions = (interaction) => {
+exports.interactions = (client, interaction) => {
   if(interaction.isCommand()) {
     const { commandName, options } = interaction;
 
@@ -27,6 +27,19 @@ exports.interactions = (interaction) => {
     if(interaction.customId == "playagain"){
       wordle.newGame(null, interaction);
     }
+  }else if(interaction.isSelectMenu()){
+    if(interaction.customId == "stattimeline"){
+      let selection = interaction.values[0];
+      var interval = null;
+      if(selection == 'lastweek'){
+        interval = "7 DAY";
+      }else if(selection == 'today'){
+        interval = "1 DAY";
+      }
+      var userID = interaction.message.embeds[0].description.split('>')[0].split('<@')[1];
+      var user = client.users.fetch(userID);
+      wordle.stats(user.User, interaction, interval);
+    }
   }
 }
 
@@ -34,9 +47,9 @@ exports.stats = new MessageActionRow()
 .addComponents(
   new MessageSelectMenu()
   .setCustomId('stattimeline')
-  .setPlaceholder('Lifetime')
+  .setPlaceholder('Select Statistic Timespan')
   .addOptions([
-    {label: 'Today', description: 'The stats of this user from today.', value: 'today'},
+    {label: 'Last 24h', description: 'The stats of this user from the last 24 hours.', value: 'today'},
     {label: 'Last Week', description: 'The stats of this user since last week.', value: 'lastweek'},
     {label: 'Lifetime', description: 'The lifetime stats of this user.', value: 'lifetime'}
   ])