+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Release Alpha 1.4.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.2) - 09.02.2022
 ### Added
 - Selection menu for stats in new module `./src/modules/interactions.js`
+- Working selection menu
 
 ### Changed
 - All interaction stuff and moved into module