+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog (Alpha Versions)
+## [Alpha 1.4.3](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.3) - 09.02.2022
+### Fixed
+- Embed lifetime selection sending reply instead of update
+
 ## [Alpha 1.4.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.2) - 09.02.2022
 ### Added
 - Selection menu for stats in new module `./src/modules/interactions.js`