+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Alpha 1.4.3](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.3) - 09.02.2022
 ### Added
 - First and last game stats
+- Stats for each server in db image
 
 ### Fixed
 - Embed lifetime selection sending reply instead of update